django_shop/example/requirements.txt
aldryn-boilerplates==0.8.0
aldryn-forms==3.0.4
awsebcli==3.14.6
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.3
blessed==1.15.0
botocore==1.12.55
bs4==0.0.1
cached-property==1.5.1
cement==2.8.2
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
cssselect==1.0.3
cssutils==1.0.2
defusedxml==0.5.0
Django==1.11.16
django-absolute==0.3
django-admin-sortable2==0.6.19
django-allauth==0.35.0
django-angular==2.0.3
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.5
django-compressor==2.2
django-emailit==0.2.3
django-embed-video==1.2.0
django-filer==1.3.2
django-filter==1.1.0
django-formtools==2.1
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-haystack==2.5.0
django-ipware==1.1.1
django-mptt==0.8.7
django-parler==1.8
django-polymorphic==1.3
django-post-office==3.0.3
django-ranged-response==0.2.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-auth==0.9.2
django-sass-processor==0.6
django-sekizai==0.10.0
django-select2==6.0.3
-e git+https://github.com/awesto/django-shop@b57b258d0cbbb96e6062c407f4f436c3a143af7d#egg=django_shop
django-simple-captcha==0.5.9
django-sizefield==0.9.1
django-sslserver==0.20
django-tablib==3.2
django-treebeard==4.2.0
djangocms-admin-style==1.2.9
djangocms-attributes-field==0.4.0
djangocms-bootstrap3==0.3.2
djangocms-cascade==0.16
djangocms-text-ckeditor==3.4.0
djangocms-video==2.1.0
djangorestframework==3.7.7
djangoshop-stripe==0.12
docker==3.6.0
docker-compose==1.21.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
drf-haystack==1.6.1
easy-thumbnails==2.5
elasticsearch==1.7.0
enum34==1.1.6
et-xmlfile==1.0.1
functools32==3.2.3.post2
hiredis==0.2.0
html5lib==0.9999999
idna==2.6
ipaddress==1.0.22
jdcal==1.4
jmespath==0.9.3
jsonfield==1.0.3
jsonschema==2.6.0
libsass==0.12.3
lxml==4.2.5
oauthlib==1.0.3
odfpy==1.4.0
olefile==0.46
openpyxl==2.4.9
pathspec==0.5.5
Pillow==4.3.0
premailer==3.2.0
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2015.7
PyYAML==3.13
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.5.0
rjsmin==1.0.12
semantic-version==2.5.0
six==1.11.0
stripe==1.53.0
tablib==0.12.1
termcolor==1.1.0
texttable==0.9.1
unicodecsv==0.14.1
Unidecode==0.4.18
urllib3==1.22
virtualenv==13.1.2
wcwidth==0.1.7
websocket-client==0.54.0
xlrd==1.1.0
xlwt==1.3.0
YURL==0.13
